requirements.txt
psycopg2-binary
pandas
numpy
nbformat
nbconvert
paramiko
scp
pillow
imageio
matplotlib
seaborn
statsmodels
openpyxl
rpy2
dvc
dvc-s3
pyodbc
sqlalchemy
